requirements/requirements-dev.txt
# This file was autogenerated by uv via the following command:
#    uv pip compile requirements/requirements-dev.in -o requirements/requirements-dev.txt
asgiref==3.8.1
    # via
    #   -r requirements/../requirements.txt
    #   django
django==5.0.6
    # via
    #   -r requirements/../requirements.txt
    #   djangorestframework
    #   djangorestframework-simplejwt
djangorestframework==3.15.1
    # via
    #   -r requirements/../requirements.txt
    #   djangorestframework-simplejwt
djangorestframework-simplejwt==5.3.1
    # via -r requirements/requirements-dev.in
iniconfig==2.0.0
    # via pytest
packaging==24.1
    # via pytest
pluggy==1.5.0
    # via pytest
pyjwt==2.9.0
    # via djangorestframework-simplejwt
pytest==8.3.2
    # via
    #   -r requirements/requirements-dev.in
    #   pytest-django
pytest-django==4.8.0
    # via -r requirements/requirements-dev.in
sqlparse==0.5.0
    # via
    #   -r requirements/../requirements.txt
    #   django
